+++ src/components/ui/index.ts
@@ -7,3 +7,4 @@ export { Loader, LoadingOverlay, Skeleton, CardSkeleton } from './Loader';
 export { UnsupportedFeature, UnsupportedBadge } from './UnsupportedFeature';
 export { UserMenu } from './UserMenu';
 export { LanguageSwitcher } from './LanguageSwitcher';
+export { VendorIcon } from './VendorIcon';